Rewrite document converter.
Use data in summary section, remove function child summary chart.
